requirements/requirements.txt
# Django Framework
Django==4.2.5

# Crispy forms : to generate form with BS5 classes
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Channels : communication channel for asynchronous operations and asynchronous protocols support
channels==4.0.0
channels-redis==4.1.0

# Tasks scheduler : useful for long running processes
django-rq==2.8.1
rq==1.15.1

# ASGI Server, will override runserver command to allow asynchronous execution
daphne==4.0.0

# Redis wrapper
redis==5.0.0

# Use IAs hosted on replicate.com
replicate==0.12.0

# HTTP Requests for humans
requests==2.31.0

# Video and audio conversions :
moviepy==1.0.3
python-ffmpeg==2.0.4

# Environnement variables loading :
python-dotenv==1.0.0

# CORS :
django-cors-headers==4.2.0

# Time zone management :
pytz==2023.3.post1

# File type discovering :
python-magic==0.4.27

# New
psycopg2==2.9.9
pgvector==0.2.5
djangorestframework==3.14.0

# Ollama Library for the API
ollama==0.1.6

# WSGI server :
gunicorn==21.2.0

# Yaml parser module :
PyYAML==6.0.1

# Allow to use color in shell :
termcolor==2.4.0

# Simple disk cache :
diskcache==5.6.3

# unicode decoding
Unidecode==1.3.8

# Structured LLM Outputs handling
instructor==1.3.5